backend/requirements.txt
faiss-cpu==1.10.0
fastapi==0.115.8
numpy==2.2.2
openai==0.28.0
pdfplumber==0.11.5
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
SQLAlchemy==2.0.38
starlette==0.45.3
tiktoken==0.8.0
tqdm==4.67.1
transformers==4.48.3
uvicorn==0.34.0
